+++ tests/core/test_smoothgrad.py
@@ -34,6 +34,6 @@ def test_get_averaged_gradients(random_data, convolutional_model):
         images, convolutional_model, 0, num_samples=num_samples
     )
 
-    expected_output_shape = (images.shape[0] / num_samples, *images.shape[1:])
+    expected_output_shape = (int(images.shape[0] / num_samples), *images.shape[1:])
 
     assert gradients.shape == expected_output_shape
